Revert "update swift int header data type (again)"
This reverts commit 02eb670065ebac813c58e72557fe7081ff41d72d.

+++ simulator/ns-3.39/src/network/utils/int-header.h
@@ -155,9 +155,9 @@ class IntHeader
         struct
         {
             // remote queueing delay
-            uint64_t remote_delay;
+            uint32_t remote_delay;
             // sent timestamp
-            uint64_t ts;
+            uint32_t ts;
             // hop count
             // required no internal padding, so outside padding?
             uint32_t nhop;
